Removes apply, allows map[?Ok] alt. return type
Apply was used to allow the caller to modify the Ok value of a Result &
Replace it with an entirely new type, but returned an unknown type. This
has been removed in favor of allowing the caller of map & mapOk to
specify a new Ok value type. If a new type is not specified, it will
default to the existing Ok value type.