Adds component to indicate that a recipe is WIP
A consumer can add this anywhere to indicate that something is a work in
progress. It renders [WIP] with a help icon to encourage users to mouse
over the acronym to read the tool-tip if they don't know what WIP stands
for. Intended to be consumed where a recipe title is used if the recipe
doesn't have the published tag.