Using useQueryParams hook in List for open Cards
Switching to use useQueryParams hook instead of the more simple & naive
implementation of native URLSearchParams. This avoids bugs where any
other queries are erased on changes to the search params used in List.

Additionally, moves logic for determining which card is open to a custom
handler on List that is passed to a Card to be called in the onClick
handler on the <a> tag. This replaces React Router's Link because a
simple navigation to `/?{cardId}` causes any active query params to be
eliminated & doesn't work on `/all-recipes`. Also changes the actual
href attribute to point to the hard `/recipe/{id}` route so that it can
still be opened in new tabs or copied as a normal link would be.